requirements.txt
Django==3.2.23
djangorestframework==3.12.4
requests==2.28.1
python-decouple==5.1